Prolong timeouts
